pyproject.toml
[tool.poetry]
name = "Arty Farty"
version = "0.1.0"
description = "Cloud AI Art Generator"
authors = ["Louis King <jinglemansweep@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.9"
dynaconf = "^3.1.12"
requests = "^2.28.2"
types-requests = "^2.28.11.15"
replicate = "^0.5.2"
rake-nltk = "^1.0.6"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
